Moved XRInteraction Editor scripts to .Editor namespace
