+++ Scripts/Editor/XRInputNameListPropertyDrawer.cs
@@ -8,7 +8,7 @@ using UnityEditor;
 using UnityEditorInternal;
 using UnityEngine;
 
-namespace Fjord.XRInteraction
+namespace Fjord.XRInteraction.Editor
 {
     [CustomPropertyDrawer(typeof(XRInputNameList))]
     public class XRInputNameListPropertyDrawer : PropertyDrawer
